fix train and val in supervised mode without available_column_item

+++ src/mars_gym/evaluation/metrics/rank.py
@@ -308,9 +308,9 @@ def personalization(predicted: List[list]) -> float:
         df = df[["index", "item"]].pivot(index="index", columns="item", values="item")
         df = df.mask(pd.notna(df), 1)
         df = df.mask(pd.isna(df), 0)
-        rec_matrix = sp.csr_matrix(df.values)
+        rec_matrix = sp.csr_matrix(df.values.astype(int))
         return rec_matrix
-
+    
     # create matrix for recommendations
     predicted = np.array(predicted)
     rec_matrix_sparse = make_rec_matrix(predicted)
